Create new file for ItemsListItem
